Adding a select players page
Added a selected player page.
On this page users can build their team and choose the players to be in their team.
The new page is connected to the database,
when a user selects players the players are saved in the database.

In addition, add several lines of code in the registration script.